refactor: création de sms-service (separation des responsabilités)
